requirements.txt
telethon
python-dotenv
google-generativeai
google-cloud-bigquery